fix(click_count): Resolve issue where anchor tags caused clicks to be counted twice
`turbolinks` is included in rails by default, unfortunately it caused a duplicate request
to be sent (identically) to the href target. This is fine for typical links, but not for links
registering clicks.

Helpers were added (`link_to_noturbo` and `link_to_short`) that set the `data-turbolinks` attribute to
false by default, disabling turbolinks for links created by those helpers.